moved models for redux and saga into separate directory, added no payload to store dispatch types
